AOSP 4.4.4_r1 check in

+++ libcutils/klog.c
@@ -35,8 +35,13 @@ void klog_set_level(int level) {
 void klog_init(void)
 {
     static const char *name = "/dev/__kmsg__";
+
+    if (klog_fd >= 0) return; /* Already initialized */
+
     if (mknod(name, S_IFCHR | 0600, (1 << 8) | 11) == 0) {
         klog_fd = open(name, O_WRONLY);
+        if (klog_fd < 0)
+                return;
         fcntl(klog_fd, F_SETFD, FD_CLOEXEC);
         unlink(name);
     }
@@ -50,6 +55,7 @@ void klog_write(int level, const char *fmt, ...)
     va_list ap;
 
     if (level > klog_level) return;
+    if (klog_fd < 0) klog_init();
     if (klog_fd < 0) return;
 
     va_start(ap, fmt);
